Partial work on render-logic thread separation
However, is does not have a bright future. The main idea behind having
separate render/logic threads would be to have the logic thread produce
ticks which would be interpolated to produce frames. However, the
interpolation is non-trivial in the presence of portals. In particular,
if the camera crosses a portal in the middle of a tick, frames before
the portal is crossed are fundamentally different from frames after
crossing the portal.

A similar problem occurs when entities cross portals. If an entity
crosses a portal quickly enough, it must be rendered on both sides of
the portal. However, interpolation may result in the entity being
rendered in its "old position" but in the "new dimension", placing it
"behind the portal" as if it had gone directly from behind the portal to
the front of the portal without changing dimensions.

Both problems stem from the fact that position during a tick is
non-linear, and are instead non-continuous piecewise linear functions.
The second issue is not that severe and can be ignored, and the first
issue could be solved by rendering a tick multiple times if the camera
crosses portals. In fact, rendering it as many times as portals are
crossed in the tick.

The problem seems too hard for too little benefit, so it is left for
later.